+++ src/logic/use-new-rental-stations/types.ts
@@ -1,4 +1,5 @@
 import { array, Infer, is, number, string, type } from 'superstruct'
+import { getTranslation } from '../../utils/utils'
 
 const RentalStationStruct = type({
     id: string(),
@@ -18,8 +19,16 @@ const RentalStationStruct = type({
 
 type RentalStation = Infer<typeof RentalStationStruct>
 
-const isRentalStation = (obj: unknown): RentalStation | null =>
+const toRentalStation = (obj: unknown): RentalStation | null =>
     is(obj, RentalStationStruct) ? obj : null
 
-export { isRentalStation }
+const byName = (a: RentalStation, b: RentalStation): number => {
+    const aName = getTranslation(a.name)
+    const bName = getTranslation(b.name)
+    if (!aName) return 1
+    if (!bName) return -1
+    return aName.localeCompare(bName, 'no')
+}
+
+export { toRentalStation, byName }
 export type { RentalStation }
