--- src/logic/use-rental-stations/UseRentalStations.mobility.graphql
@@ -1,40 +1,0 @@
-fragment StationFragment on Station {
-    id
-    name {
-        translation {
-            language
-            value
-        }
-    }
-    lat
-    lon
-    numBikesAvailable
-    numDocksAvailable
-}
-
-query UseRentalStations_StationsById (
-    $stationIds: [String!]!
-) @api(name: mobility) {
-    stationsById(
-        ids: $stationIds
-    ) {
-        ...StationFragment
-    }
-}
-
-
-query UseRentalStations_NearbyStations (
-    $lat: Float!
-    $lon: Float!
-    $range: Int!
-    $availableFormFactors: [FormFactor]
-) @api(name: mobility) {
-    stations(
-        lon: $lon
-        lat: $lat
-        range: $range
-        availableFormFactors: $availableFormFactors
-    ) {
-        id
-    }
-}
